VideoCommon: refactor the custom pipeline logic in the graphics mod action into a separate class, so it is reusable
